requirements.txt
ultralytics>=8.0.0
insightface>=0.7.3
onnxruntime>=1.16.0
opencv-python>=4.8.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
boxmot>=10.0.0
scipy>=1.11.0
flask>=3.0.0
